add support for enum types
